# swagger.yaml
swagger: '2.0'
info:
  description: Widgets - Documentation Api
  version: 1.0.0
  title: Widget API
host: 'localhost:8080'
schemes:
  - http
paths:
  /user:
    get:
      tags:
        - user
      summary: Get list of users
      operationId: getUsers
      security:
        - Bearer: []
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
  '/user{userId}':
    get:
      tags:
        - user
      summary: Find user by ID
      operationId: getUser
      security:
        - Bearer: []
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: ID of user to return
          required: true
          type: integer
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
  /widget:
    get:
      tags:
        - widget
      summary: Get list of widget
      operationId: getWidgets
      description: Return all widgets
      security:
        - Bearer: []
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Widget'
    post:
      tags:
        - widget
      summary: Create Widget
      operationId: createWidget
      security:
        - Bearer: []
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: name
          in: formData
          description: name
          required: true
          type: string
        - name: color
          in: formData
          description: color
          required: true
          type: string
        - name: price
          in: formData
          description: price
          required: true
          type: number
        - name: inventory
          in: formData
          description: inventory
          required: true
          type: integer
        - name: melts
          in: formData
          description: melts
          required: true
          type: boolean
      responses:
        default:
          description: successful operation
          schema:
            $ref: '#/definitions/Widget'
  '/widget{widgetId}':
    get:
      tags:
        - widget
      summary: Find widget by ID
      operationId: getWidget
      security:
        - Bearer: []
      produces:
        - application/json
      parameters:
        - name: widgetId
          in: path
          description: ID of widget to return
          required: true
          type: integer
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Widget'
    put:
      tags:
        - widget
      summary: Update Widget
      operationId: updateWidget
      security:
        - Bearer: []
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: id
          in: formData
          description: id
          type: integer
          required: true
        - name: name
          in: formData
          description: name
          required: true
          type: string
        - name: color
          in: formData
          description: color
          required: true
          type: string
        - name: price
          in: formData
          description: price
          required: true
          type: number
        - name: inventory
          in: formData
          description: inventory
          required: true
          type: integer
        - name: melts
          in: formData
          description: melts
          required: true
          type: boolean
      responses:
        default:
          description: successful operation
          schema:
            $ref: '#/definitions/Widget'
  /token:
    post:
      tags:
        - token
      summary: Create Token
      description: Returns a token using a valid userID
      operationId: createToken
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: id
          in: formData
          description: user id
          required: true
          type: integer
      responses:
        default:
          description: successful operation
          schema:
            $ref: '#/definitions/Token'
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
definitions:
  User:
    type: object
    required:
      - id
      - name
      - gravatar
    properties:
      id:
        type: integer
        example: 23
      name:
        type: string
        example: Colin
      gravatar:
        type: string
        example: >-
          http://www.gravatar.com/avatar/a51972ea936bc3b841350caef34ea47e?s=64&d=monsterid
    xml:
      name: User
  Widget:
    type: object
    required:
      - id
      - name
      - color
      - price
      - inventory
      - melts
    properties:
      id:
        type: integer
        example: 16
      name:
        type: string
        example: Claster
      color:
        type: string
        example: off-white
      price:
        type: number
        example: 12.56
      inventory:
        type: integer
        example: 18
      melts:
        type: boolean
        example: true
    xml:
      name: Pet
  Token:
    type: object
    required:
      - token
    properties:
      token:
        type: string
        example: >-
          eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE1MDg3NzUwMDksInVzZXIiOnsiaWQiOjIsIm5hbWUiOiJLeWxlIiwiZ3JhdmF0YXIiOiJodHRwOi8vd3d3LmdyYXZhdGFyLmNvbS9hdmF0YXIvNDMyZjNlMzUzYzY4OWZjMzdhZjg2YWU4NjFkOTM0Zjk_cz02NFx1MDAyNmQ9bW9uc3RlcmlkIn19.Eyx3RfvRZnxjJXSHp_tichZcS4eETXckFib-9iNN8a8